move _extend_population method to PopulationalOptimiser
